+++ src/main/java/kitchenpos/menu/dto/ProductRequest.java
@@ -1,7 +1,7 @@
-package kitchenpos.product.dto;
+package kitchenpos.menu.dto;
 
 import java.math.BigDecimal;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 public class ProductRequest {
     private String name;
